go-1270: add enabledtabs option in config
Add an `enabledTabs` option to the configuration file, allowing a
string[] to be passed to the application to define which tabs should be
displayed.

This functionality allows us to decouple OpenPedigree from a particular
application or required data set.

Potentially, a future version should look to separate the
personal/clinical data collection code from the application itself.
This would allow application-specific functionality to be included
within OpenPedigree.  However, this is rather beyond the scope that I'm
currently working under!