go-1992: create add/remove cohort member fns
Update `app.js` to include functions to add a new member to a cohort,
and remove an existing member from a cohort.

These functions can then be triggered on the relevant events.  The
functions have been reused from the existing Gen-o codebase, but with
the name altered so that they can be allowed to diverge over time if
required.