gh-737: add val for 0 children in dynamicgraph.js
Update `dynamicGraph.js` to provide a returned value if there are no
children, as this seems to be where the error occurs - it had used just
`return;`, which would come up as an undefined value.

Given this information, it may be a possible cause of gh-688.  This has
been noted in the ticket, and will be tested.